+++ extensions/common/azure/azure-eventgrid/src/main/java/org/eclipse/edc/azure/event/AzureEventGridPublisher.java
@@ -17,9 +17,9 @@ package org.eclipse.edc.azure.event;
 import com.azure.core.util.BinaryData;
 import com.azure.messaging.eventgrid.EventGridEvent;
 import com.azure.messaging.eventgrid.EventGridPublisherAsyncClient;
-import org.eclipse.edc.connector.transfer.spi.observe.TransferProcessListener;
-import org.eclipse.edc.connector.transfer.spi.types.TransferProcess;
-import org.eclipse.edc.connector.transfer.spi.types.TransferProcessStates;
+import org.eclipse.edc.connector.controlplane.transfer.spi.observe.TransferProcessListener;
+import org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferProcess;
+import org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferProcessStates;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.jetbrains.annotations.NotNull;
 import reactor.core.publisher.BaseSubscriber;
