+++ extensions/control-plane/provision/provision-blob/src/main/java/org/eclipse/edc/connector/provision/azure/blob/ObjectStorageResourceDefinition.java
@@ -15,7 +15,8 @@
 
 package org.eclipse.edc.connector.provision.azure.blob;
 
-import org.eclipse.edc.connector.transfer.spi.types.ResourceDefinition;
+
+import org.eclipse.edc.connector.controlplane.transfer.spi.types.ResourceDefinition;
 
 import java.util.Objects;
